test: extract delegated skill invocation suite
